ducati: fix up some nasty (insane?) build issues
in XDC, when a package has a 'requires' statement, it implies that
when running xdc -PD <target> it would potentially 'touch' the
dependency as well. E.g. when running xdc -PD clean, it would end
up cleaning the dependent package... even if that looks cumbersome.

as a consequence it is recommended to avoid 'requiring' packages
outside of our own tree. especially in the case of tools such as
XDC, BIOS, CE, ... which are typically installed in a 'read only'
central server.

so this commit removes the 'requires' and fixes the resulting
link issues.

Signed-off-by: Nicolas Dechesne <n-dechesne@ti.com>